[Optiq-Systems] Compare: share the A/B pane chrome and route requests per source
Move the split, the card and the source title of the side by side views
into rocprofvis_compare_panes, so the analysis tables, top events, events
and track details all build their panes the same way instead of each one
hand rolling it.

Give every source its own client request id and TablesModel slot, and keep
requests for one controller table type serialized. A fetch that finds the
table busy is parked by InfiniteScrollTable and reissued from Update(),
which replaces the separate selection retry MultiTrackTable used to keep.

+++ src/view/src/rocprofvis_track_details.h
@@ -66,9 +66,11 @@ private:
     void RenderTable(InfoTable& table, const char* table_id,
                      const AnalysisTrackStatistics* = nullptr);
 
-    // Renders track detail cards, filtered to source_index in compare mode; returns count.
-    int  RenderDetailList(std::optional<uint64_t> source_index);
-    void RenderSourceColumn(size_t source_index);
+    /* Renders the selected tracks' cards and returns how many were drawn. In
+     * compare mode source_index limits a column to one source's tracks.
+     */
+    size_t RenderDetailList(std::optional<uint64_t> source_index);
+    void   RenderSourceColumn(size_t source_index);
 
     std::shared_ptr<TrackTopology>     m_track_topology;
     DataProvider&                      m_data_provider;
